+++ src/api/auth.js
@@ -1,10 +1,21 @@
 import instanceBaza2 from './config/instance-baza2';
+// Управління access_token
+export const token = {
+  get:()=>{
+    const token = sessionStorage.getItem('access_token');
+    return token
+  },
 
-const token = {
   set: (token) => {
+    sessionStorage.setItem(
+      'access_token',
+      token
+    )
     instanceBaza2.defaults.headers.common.Authorization = `${token}`;
   },
+
   reset: () => {
+    sessionStorage.removeItem('access_token')
     instanceBaza2.defaults.headers.common.Authorization = '';
   },
 };
