+++ src/components/admin-page/LayoutAdmin/LayoutAdmin.module.scss
@@ -2,5 +2,5 @@
   display: flex;
   flex: 1 100%;
   max-height: 100vh;
-  min-width: 1000px;
+  //min-width: 1000px;
 }
