+++ src/components/admin-page/NavBar/NavBar.module.scss
@@ -64,7 +64,7 @@
     //float: right;
     margin-left: auto;
     margin-bottom: 20px;
-
+    height: 40px;
     border-radius: 4px 0 0 4px;
     padding: 13px 16px;
     & .icon {
