--- src/components/login-page/ForgotPassword/ForgotPassword.module.scss
@@ -1,167 +1,0 @@
-@use "../../../styles/mixins" as mix;
-@use "../../../styles/variables" as var;
-
-.section{
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-height: 100vh;
- 
-  padding: 10px;
-  @media (min-height: 1024px) {
-    padding: 50px 16px;
-  }
-
-  .form{
-    border-radius: 8px;
-    background: #d2d2d2;
-    color: #020202;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 18px;
-    padding: 50px;
-    max-width: 600px;
-
-    @media (min-height: 1024px) {
-      padding: 50px;
-      gap: 36px;
-    }
-
-    .title{
-      text-align: center;
-      & > h2{
-        @include mix.apply-font-styles(var.$h1-font);
-        font-weight: 600;
-        font-size: 34px;
-
-        @media (min-width: 1600px) {
-          font-size: 40px;
-        }
-      }
-  
-      & > p{
-        @include mix.apply-font-styles(var.$h2-font);
-        font-weight: 400;
-        font-size: 20px;
-      }
-    }
-
-    // & > a {
-    //   @include mix.apply-font-styles(var.$h2-font);
-    //   font-weight: 400;
-    //   font-size: 20px;
-    //   background: var.$icons-gradient;
-    //   background-clip: text;
-    //     -webkit-background-clip: text;
-    //     -webkit-text-fill-color: transparent;
-    //   transition: all .3s;
-    // }
-    // & > a:hover{
-    //   opacity: .7;
-    // }
-
-    .list {
-      display: flex;
-      flex-direction: column;
-      //gap: 20px;
-      gap: 32px;
-
-      @media (min-height: 1024px) {
-        // height: clamp(550px, 86vh, 860px);
-        gap: 32px;
-      }
-
-      // .list_item{
-      //   position: relative;
-      //   .btn{
-      //     position: absolute;
-      //     right: 10px;
-      //     top: 40px;
-      //   }
-      // }
-      .btns{
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        //background-color: aqua;
-        //width: 100%;
-        //gap: 10px;
-        & > button{
-          height: 55px;
-          font-weight: 600;
-          font-size: 18px;
-        }
-        .btn_cancel{
-          border: 3px solid  #2e0535;
-          box-shadow: inset 3px 3px 4px 0 rgba(0, 0, 0, 0.25);
-          color: #020202;
-        }
-      }
-      .item {
-        width: clamp(280px, 84vw, 360px);
-        display: flex;
-        flex-direction: column;
-        position: relative;
-        & > input{
-          padding: 10px 35px 10px 10px;
-        }
-        & > label{
-          font-weight: 600;
-          font-size: 18px;
-        }
-
-        // .btn_option{
-        //   display: flex;
-        //   align-items: center;
-        //   cursor: pointer;
-        //   gap: 12px;
-        //   @include mix.apply-font-styles(var.$h3-font);
-        //   font-weight: 600;
-        //   font-size: 18px;
-  
-        //   &.agree{
-        //     @include mix.apply-font-styles(var.$text-16-font);
-        //     font-size: 16px;
-        //   }
-  
-        //   & input{
-        //     display: none;
-        //   }
-  
-        //   .check{
-        //     width: 24px;
-        //     height: 24px;
-        //     border-radius: 2px;
-        //     border: 1px solid var.$secondary-gray;
-           
-        //     & svg{
-        //       opacity: 0;
-        //       display: block;
-        //       border-radius: 2px;
-        //       width: 24px;
-        //       height: 24px;
-        //     }
-  
-        //     &._active{
-        //       border: none;
-  
-        //       & svg{
-        //         opacity: 1;
-        //       }
-        //     }
-        //   }
-        // }
-
-
-
-      }
-   
-      .error_modal{
-        @include mix.apply-font-styles(var.$text-16-font);
-        font-size: 14px;
-        color: var.$secondary-error;
-      }
-    }
-  }
-}
