+++ src/components/login-page/ForgotPassword/FormForgotPassword/FormForgotPassword.module.scss
@@ -1,0 +1,47 @@
+@use "../../../../styles/mixins" as mix;
+@use "../../../../styles/variables" as var;
+
+.form{
+  .list {
+    display: flex;
+    flex-direction: column;
+    gap: 32px;
+
+    .btns{
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
+
+      & > button{
+        height: 55px;
+        font-weight: 600;
+        font-size: 18px;
+      }
+      .btn_cancel{
+        padding: 15px 24px;
+        border: 3px solid  #2e0535;
+        box-shadow: inset 3px 3px 4px 0 rgba(0, 0, 0, 0.25);
+        color: #020202;
+      }
+    }
+    .item {
+      width: clamp(280px, 84vw, 360px);
+      display: flex;
+      flex-direction: column;
+      position: relative;
+      & > input{
+        padding: 10px 35px 10px 10px;
+      }
+      & > label{
+        font-weight: 600;
+        font-size: 18px;
+      }
+    }
+  
+    .error_modal{
+      @include mix.apply-font-styles(var.$text-16-font);
+      font-size: 14px;
+      color: var.$secondary-error;
+    }
+  }
+}
